key unsigned long

+++ lib_lem_in/free_hashtable.c
@@ -6,7 +6,7 @@
 /*   By: vkurkela <vkurkela@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/19 20:06:25 by vkurkela          #+#    #+#             */
-/*   Updated: 2020/04/04 13:53:57 by vkurkela         ###   ########.fr       */
+/*   Updated: 2020/04/04 14:00:34 by vkurkela         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,7 +44,7 @@ static void	free_hashes(t_hashtable *root)
 
 void		free_hashtable(t_lem_in *anthill)
 {
-	int key;
+	unsigned long	key;
 
 	key = 0;
 	if (!anthill->hashtable)
